# Microoptimize the rewriter class