```
                           31f7289806254484c8:  760925b964ace749eb:
      django_template3.py             3.0s (4)             3.0s (4)  -2.6%
            pyxl_bench.py             2.6s (4)             2.6s (4)  -1.3%
sqlalchemy_imperative2.py             3.1s (4)             3.0s (4)  -1.7%
                  geomean                 2.9s                 2.8s  -1.9%
```